+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -   View details for Peaks
 -   Adds view sensor button to results table which opens dialog listing sensors; includes button to site page
 -   Makes siteid a link to site page
+-   Event populated in map filter form on page load
+-   Map filters after navigating from map page
 
 ### Changed
 
@@ -29,6 +31,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 -  Sensor and HWM files not always displaying in peak dialogs when event selected
 -  Sensor markers not displaying on map when no deployment type
 -  View Details/Edit buttons displayed in both column and menu
+-  Site icon not displayed on map on site pages when no sensors or hwms
+-  Surveyed RP Only Sites not send with map filter query
+-  Markers not showing up after returning to map page
 
 ## [v0.6.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.6.0) - 2022-01-25